Refactor into acceptor module
