+++ tests/core/test_acceptor.py
@@ -33,7 +33,7 @@ class TestAcceptor(unittest.TestCase):
 
     @mock.patch('selectors.DefaultSelector')
     @mock.patch('socket.fromfd')
-    @mock.patch('proxy.core.acceptor.recv_handle')
+    @mock.patch('proxy.core.acceptor.acceptor.recv_handle')
     def test_continues_when_no_events(
             self,
             mock_recv_handle: mock.Mock,
@@ -57,7 +57,7 @@ class TestAcceptor(unittest.TestCase):
     @mock.patch('threading.Thread')
     @mock.patch('selectors.DefaultSelector')
     @mock.patch('socket.fromfd')
-    @mock.patch('proxy.core.acceptor.recv_handle')
+    @mock.patch('proxy.core.acceptor.acceptor.recv_handle')
     def test_accepts_client_from_server_socket(
             self,
             mock_recv_handle: mock.Mock,
